Move Controller Report Base
Moved controller report base to extended controller lib directory.
Add class documentation.
